+++ plan.md
@@ -1,1 +1,4 @@
-`src/shared/ui/` の中から特定のページでしか使われていないコンポーネントを `src/pages/〇〇/ui` へ移動する。
+✅ `src/shared/ui/FollowButton` を `src/features/follow/ui` へ移動する。
+
+✅ `src/shared/ui/UsersList` を `src/widgets/user-list` へ移動する
+- [ ] `
