+++ src/features/follow/ui/index.ts
@@ -1,0 +1,1 @@
+export { FollowButton } from "./FollowButton";
