+++ src/pages/search/ui/Search.tsx
@@ -8,7 +8,8 @@ import useSWR from "swr";
 import { useSearchPosts } from "@/shared/api";
 import type { User } from "@/shared/entities";
 import { fetcher } from "@/shared/lib";
-import { PostsList, UsersList } from "@/shared/ui";
+import { PostsList } from "@/shared/ui";
+import { UsersList } from "@/widgets/user-list";
 
 interface SearchProps {
 	currentUser: User;
