+++ src/widgets/user-list/index.ts
@@ -1,0 +1,1 @@
+export { UsersList, UserItem } from "./ui";
