+++ src/widgets/user-list/ui/index.ts
@@ -1,0 +1,2 @@
+export { UsersList } from "./UsersList";
+export { UserItem } from "./UserItem";
